refactor: move helper functions to its' own file
