+++ sumOfOther/sumOfOther.js
@@ -1,21 +1,7 @@
-function getArraySum(arr) {
-  const reducer = (accumulator, currentValue) => accumulator + currentValue;
-  return arr.reduce(reducer);
-}
-
-function getOtherItems(arr, index) {
-  return arr.filter((item, i) => i !== index);
-}
-
-function getSumOfOtherForItem(arr, index) {
-  const otherItems = getOtherItems(arr, index);
-  return getArraySum(otherItems);
-}
+const { getSumOfOtherForItem } = require('./helpers');
 
 function sumOfOther(arr) {
   return arr.map((item, index) => getSumOfOtherForItem(arr, index));
 }
 
-module.exports = {
-  sumOfOther, getSumOfOtherForItem, getOtherItems, getArraySum,
-};
+module.exports = sumOfOther;
