+++ sumOfOther/sumOfOther.test.js
@@ -1,7 +1,6 @@
 const assert = require('assert');
-const {
-  sumOfOther, getSumOfOtherForItem, getOtherItems, getArraySum,
-} = require('./sumOfOther');
+const sumOfOther = require('./sumOfOther');
+const { getArraySum, getOtherItems, getSumOfOtherForItem } = require('./helpers');
 
 describe('sumOfOther', function () {
   it('should sum basic numbers', function () {
@@ -15,47 +14,47 @@ describe('sumOfOther', function () {
   });
 });
 
-describe('getSumOfOtherForItem', function () {
-  it('1', function () {
-    assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 0), 8);
+describe('helpers', function () {
+  describe('getSumOfOtherForItem', function () {
+    it('1', function () {
+      assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 0), 8);
+    });
+    it('2', function () {
+      assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 1), 7);
+    });
+    it('3', function () {
+      assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 2), 6);
+    });
+    it('4', function () {
+      assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 3), 9);
+    });
   });
-  it('2', function () {
-    assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 1), 7);
+  describe('getOtherItems', function () {
+    it('1', function () {
+      assert.deepEqual(getOtherItems([2, 3, 4, 1], 0), [3, 4, 1]);
+    });
+    it('2', function () {
+      assert.deepEqual(getOtherItems([2, 3, 4, 1], 1), [2, 4, 1]);
+    });
+    it('3', function () {
+      assert.deepEqual(getOtherItems([2, 3, 4, 1], 2), [2, 3, 1]);
+    });
+    it('4', function () {
+      assert.deepEqual(getOtherItems([2, 3, 4, 1], 3), [2, 3, 4]);
+    });
   });
-  it('3', function () {
-    assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 2), 6);
-  });
-  it('4', function () {
-    assert.deepEqual(getSumOfOtherForItem([2, 3, 4, 1], 3), 9);
-  });
-});
-
-describe('getOtherItems', function () {
-  it('1', function () {
-    assert.deepEqual(getOtherItems([2, 3, 4, 1], 0), [3, 4, 1]);
-  });
-  it('2', function () {
-    assert.deepEqual(getOtherItems([2, 3, 4, 1], 1), [2, 4, 1]);
-  });
-  it('3', function () {
-    assert.deepEqual(getOtherItems([2, 3, 4, 1], 2), [2, 3, 1]);
-  });
-  it('4', function () {
-    assert.deepEqual(getOtherItems([2, 3, 4, 1], 3), [2, 3, 4]);
-  });
-});
-
-describe('getArraySum', function () {
-  it('1', function () {
-    assert.equal(getArraySum([2, 3, 4, 1]), 10);
-  });
-  it('2', function () {
-    assert.equal(getArraySum([2, 4, 1]), 7);
-  });
-  it('3', function () {
-    assert.equal(getArraySum([2, 3, 1]), 6);
-  });
-  it('4', function () {
-    assert.equal(getArraySum([2, 3, 4]), 9);
+  describe('getArraySum', function () {
+    it('1', function () {
+      assert.equal(getArraySum([2, 3, 4, 1]), 10);
+    });
+    it('2', function () {
+      assert.equal(getArraySum([2, 4, 1]), 7);
+    });
+    it('3', function () {
+      assert.equal(getArraySum([2, 3, 1]), 6);
+    });
+    it('4', function () {
+      assert.equal(getArraySum([2, 3, 4]), 9);
+    });
   });
 });
